style: css 코드 및 상수 데이터 외부 파일에 저장
