+++ week3/src/Main/Card/CardData.jsx
@@ -1,0 +1,13 @@
+const cardData = [
+    {id: 1, backImg: 'src/assets/back1.png'},
+    {id: 2, backImg: 'src/assets/back2.png'},
+    {id: 3, backImg: 'src/assets/back3.png'},
+    {id: 4, backImg: 'src/assets/back4.png'},
+    {id: 5, backImg: 'src/assets/back5.png'},
+    {id: 6, backImg: 'src/assets/back6.png'},
+    {id: 7, backImg: 'src/assets/back7.png'},
+    {id: 8, backImg: 'src/assets/back8.png'},
+    {id: 9, backImg: 'src/assets/back9.png'}
+]
+
+export default cardData
