+++ week3/src/Main/LevelData.jsx
@@ -1,0 +1,7 @@
+const gameLevel = [
+    {level: 'Easy', cardNum: 5},
+    {level: 'Normal', cardNum: 7},
+    {level: 'Hard', cardNum: 9},
+]
+
+export default gameLevel
